Added building, working on builder unit
Added a proper team system, added colours for different teams. Made a system that picks up when objects are in the area. Made a new Entity base class that the building and unit classes inherit from. The player can now place buildings that the server will see, although the builder in not yet able to finish building these buildings.